packages/compliance/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rare-archive-compliance"
version = "0.1.0"
description = "aDNA validation, FAIR scoring, and governance for the Rare AI Archive"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
